added baseuserlogindto for both and also organized the dtos for auth into seperate folder for regsiter and login
