--- src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/UserLoginDTO.java
@@ -1,14 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
-import ch.uzh.ifi.hase.soprafs24.constant.UserAccountType;
-import lombok.Getter;
-import lombok.Setter;
-
-
-@Getter @Setter
-public class UserLoginDTO {
-
-    private String username;
-    private String password;
-    private UserAccountType userAccountType; // "DRIVER" or "REQUESTER"
-
-}
