+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/auth/register/DriverRegisterDTO.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
+package ch.uzh.ifi.hase.soprafs24.rest.dto.auth.register;
 
 import lombok.Getter;
 import lombok.Setter;
