+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/auth/register/RequesterRegisterDTO.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
+package ch.uzh.ifi.hase.soprafs24.rest.dto.auth.register;
 
 public class RequesterRegisterDTO extends BaseUserRegisterDTO {
