+++ src/main/java/ch/uzh/ifi/hase/soprafs24/security/RequesterRegisterationService.java
@@ -3,7 +3,7 @@ package ch.uzh.ifi.hase.soprafs24.security;
 import org.springframework.stereotype.Service;
 
 import ch.uzh.ifi.hase.soprafs24.entity.Requester;
-import ch.uzh.ifi.hase.soprafs24.rest.dto.RequesterRegisterDTO;
+import ch.uzh.ifi.hase.soprafs24.rest.dto.auth.register.RequesterRegisterDTO;
 
 @Service
 public class RequesterRegisterationService {
